Extract package && apply spotless
